+++ scraper.js
@@ -1,13 +1,14 @@
 const async = require("async-q"),
   config = require("./config"),
-  eztv = require("./providers/eztv")("EZTV"),
+  eztv = require("./providers/show/eztv")("EZTV"),
+  kat = require("./providers/show/kat"),
   util = require("./util");
 
 /* Start scraping from KAT. */
 const scrapeKAT = () => {
   return async.eachSeries(config.providers, (provider) => {
     util.setStatus("Scraping " + provider.name);
-    const kat = require("./providers/kat")(provider.name);
+    const katProvider = kat(provider.name);
     return util.spawn(kat.search(provider)).then((response) => {
       console.log(provider.name + ": Done.");
       return response;