+++ setup.js
@@ -7,13 +7,13 @@ const bodyParser = require("body-parser"),
   config = require("./config");
 
 const mongoose = require("mongoose");
-mongoose.Promise = require("q").Promise;
+mongoose.Promise = global.Promise;
 
 RegExp.escape = (text) => {
   return text.replace(/[-[\]{}()*+?.,\\^$|#\s]/g, "\\$&");
 };
 
-mongoose.connect("mongodb://" + config.dbHosts.join(",") + "/popcorn_shows", {
+mongoose.connect("mongodb://" + config.dbHosts.join(",") + "/popcorn", {
   db: {
     native_parser: true
   },